RF: Move with_result_progress from addurls to log
This decorator is not specific to addurls.  It can be used with any
function that yields results.

This is pure code movement.
