+++ datalad/plugin/addurls.py
@@ -9,7 +9,7 @@
 """Create and update a dataset from a list of URLs.
 """
 
-from collections import defaultdict, Mapping
+from collections import Mapping
 from functools import partial
 from itertools import dropwhile
 import logging
@@ -21,16 +21,15 @@ from six import string_types
 from six.moves.urllib.parse import urlparse
 
 from datalad.dochelpers import exc_str
-from datalad.log import log_progress
+from datalad.log import log_progress, with_result_progress
 from datalad.interface.base import Interface
 from datalad.interface.base import build_doc
 from datalad.interface.results import annexjson2result, get_status_dict
 from datalad.interface.common_opts import nosave_opt
-from datalad.support import ansi_colors
 from datalad.support.exceptions import AnnexBatchCommandError
 from datalad.support.network import get_url_filename
 from datalad.support.s3 import get_versioned_url
-from datalad.utils import assure_list, optional_args
+from datalad.utils import assure_list
 
 lgr = logging.getLogger("datalad.plugin.addurls")
 
@@ -474,70 +473,6 @@ def extract(stream, input_type, url_format="{0}", filename_format="{1}",
         filename_format)
     subpaths = _format_filenames(format_filename, rows_with_url, infos)
     return infos, subpaths
-
-
-@optional_args
-def with_result_progress(fn, label="Total", unit=" Files"):
-    """Wrap a progress bar, with status counts, around a function.
-
-    Parameters
-    ----------
-    fn : generator function
-        This function should accept a collection of items as a
-        positional argument and any number of keyword arguments.  After
-        processing each item in the collection, it should yield a status
-        dict.
-    label, unit : str
-        Passed to log.log_progress.
-
-    Returns
-    -------
-    A variant of `fn` that shows a progress bar.  Note that the wrapped
-    function is not a generator function; the status dicts will be
-    returned as a list.
-    """
-    # FIXME: This emulates annexrepo.ProcessAnnexProgressIndicators.  It'd be
-    # nice to rewire things so that it could be used directly.
-
-    def count_str(count, verb, omg=False):
-        if count:
-            msg = "{:d} {}".format(count, verb)
-            if omg:
-                msg = ansi_colors.color_word(msg, ansi_colors.RED)
-            return msg
-
-    pid = str(fn)
-    base_label = label
-
-    def wrapped(items, **kwargs):
-        counts = defaultdict(int)
-
-        label = base_label
-        log_progress(lgr.info, pid,
-                     "%s: starting", label,
-                     total=len(items), label=label, unit=unit)
-
-        results = []
-        for res in fn(items, **kwargs):
-            counts[res["status"]] += 1
-            count_strs = (count_str(*args)
-                          for args in [(counts["notneeded"], "skipped", False),
-                                       (counts["error"], "failed", True)])
-            if counts["notneeded"] or counts["error"]:
-                label = "{} ({})".format(
-                    base_label,
-                    ", ".join(filter(None, count_strs)))
-
-            log_progress(
-                lgr.error if res["status"] == "error" else lgr.info,
-                pid,
-                "%s: processed result%s", base_label,
-                " for " + res["path"] if "path" in res else "",
-                label=label, update=1, increment=True)
-            results.append(res)
-        log_progress(lgr.info, pid, "%s: done", base_label)
-        return results
-    return wrapped
 
 
 @with_result_progress("Adding URLs")
